requirements.txt
analytics-python==1.2.9
appnope==0.1.0
awscli==1.14.64
backports.shutil-get-terminal-size==1.0.0
boto3==1.6.17
botocore==1.9.17
certifi==2018.1.18
chardet==3.0.4
click==6.7
colorama==0.3.7
decorator==4.2.1
docutils==0.14
enum34==1.1.6
Flask==0.12.2
Flask-WTF==0.14.2
futures==3.2.0
gunicorn==19.7.1
idna==2.6
ipdb==0.11
ipython==5.5.0
ipython-genutils==0.2.0
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
MarkupSafe==1.0
pathlib2==2.3.0
pexpect==4.4.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pyasn1==0.4.2
Pygments==2.2.0
python-dateutil==2.6.1
pytz==2018.3
PyYAML==3.12
redis==2.10.6
requests==2.18.4
rq==0.10.0
rsa==3.4.2
s3transfer==0.1.13
scandir==1.7
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
Werkzeug==0.14.1
WTForms==2.1
